+++ models/params_test.go
@@ -1,4 +1,4 @@
-package params
+package models
 
 import (
 	"net/url"
